Main2

+++ app/Http/Controllers/API/PackageController.php
@@ -55,7 +55,12 @@ class PackageController extends Controller
     public function show(Package $package)
     {
         if($package['user_id'] === auth('api')->user()->id OR $package['user_id'] === 1){
-            $package->load('question');
+            
+           $package->load(['question' => function ($query) {
+                $query->where('user_id', auth('api')->user()->id)->orwhere('user_id', 1);
+                
+            }]);
+            
             return $package;
             
         }
@@ -77,9 +82,11 @@ class PackageController extends Controller
             'name' => 'required',
             'text' => 'required',
         ]);
+        if($package['user_id'] === auth('api')->user()->id){  
         $package->update(
             $request->all()
         );
+    }
 
          
     }
@@ -92,9 +99,9 @@ class PackageController extends Controller
      */
     public function destroy(Package $package)
     {
-
+        if($package['user_id'] === auth('api')->user()->id){ 
         $package->delete();
+        }
 
-        
     }
 }
